Fix CI mypy failure structurally and clean up type-debt across cflibs
CI's `build-test-lint (3.12)` was failing on `mypy cflibs/` with exit 2,
because zarr ≥3.2 uses PEP 695 generic class syntax (`class AsyncArray[T]`)
in `zarr/core/array.py:304` while pyproject pinned `python_version = "3.10"`.
Mypy followed the import, tried to parse 3.12 grammar under 3.10 rules, and
crashed. The same failure was present on main; it just hadn't been blamed
yet.

Structural fix:

* Bump `python_version` from "3.10" → "3.11" to match `requires-python`.
* `follow_imports = "skip"` for optional deps with type-stub gaps that have
  caused this category before: zarr, numpyro, arviz, dynesty. cflibs uses
  these only through `try: import; except ImportError` shims, so skipping
  type recursion does not lose useful coverage.

While in the file I removed the global `cflibs.*` suppression that was
hiding 147 real errors, fixed everything that was tractable, and narrowed
the remaining suppression to six JAX-heavy modules with only the three
error codes that are genuinely about the JAX `Array` ↔ Python `float`
boundary. Net: 147 → 0 mypy errors with full type checking on 162 of 168
files, vs. 147 silent errors on 168 files before.

Concrete fixes (≈50):

* Optional-import `Module | None` shim pattern: 17 files, narrow
  `# type: ignore[assignment]` on the `except ImportError: x = None` line.
* `Path` vs `str` mix-up in `manifold/vector_index.py` and
  `manifold/basis_index.py` `save()`/`load()`: widen signatures to
  `str | Path`, use a separate `Path` local instead of reassigning the
  parameter. Real bugs that stayed hidden by the override.
* `vector_index.VectorIndex.index` was None-typed and used as a FAISS index;
  retype as `Any` since FAISS exposes no common base class.
* `composition_records` loop in `benchmark/unified.py` was reusing the
  `record` variable from a previous `id_records` loop, leaking the wrong
  dataclass type — rename to `comp_record`. Real (silent) bug.
* `quality_metrics: Dict[str, float]` widened to `Dict[str, Any]` in
  `inversion/solve/iterative.py`. Streaming pipeline puts strings (mode
  names, warnings, error messages) and bools (LTE flag) here intentionally.
* `_pack_observations` return type `Dict[str, np.ndarray]` widened to
  `Dict[str, Any]` since it includes the integer count `n_obs`.
* `compute_spectrum_batch` / `_grid` / `_ensemble` return types widened to
  `List[Tuple[Optional[ndarray], Optional[ndarray]]]` to reflect the
  documented `(None, None)` failure entries.
* `np.searchsorted`/`np.argmin`/`np.mean` results coerced with `int()` /
  `float()` at the boundary (pca.py, alias.py, comb.py, basis_index.py,
  metrics.py, temporal.py, self_absorption.py, cdsb.py).
* `saha_boltzmann.py:136`: assert `eff_ip_II is not None` to narrow what the
  outer `if ip_II is not None` already implies.
* `radiation/emissivity.py`: implicit Optional default → explicit `Optional[float]`,
  plus assert isinstance ndarray at the per-line broadening branch.
* `core/factory.py`: cast(Any, ...) for the SolverStrategy/InstrumentModel
  dynamic dispatch, with comment explaining why.
* `hardware/factory.py`: separate variable for the default-class lookup
  so mypy doesn't carry the non-Optional type forward; cast through Any
  to satisfy the abstract-class check.
* `inversion/preprocess/deconvolution.py`: explicit
  `import jax.scipy.optimize` so `jax.scipy.optimize.minimize` resolves.
* `inversion/identify/line_detection.py`: rename `transition` →
  `matched` after `_match_transition()` to avoid clobbering the surrounding
  loop's non-Optional `transition` variable.
* `io/exporters.py:_dataclass_to_dict`: widen return type to `Any` because
  it can return `str(obj)` when `max_depth` is exhausted.

Pre-existing JAX `Array`/float boundary in
`inversion/solve/{bayesian,joint_optimizer,coarse_to_fine}`,
`inversion/preprocess/deconvolution`, and `radiation/{profiles,stark}`
needs proper widening of function signatures from `float` to
`float | jax.Array`. That's an API-surface refactor across thousands of
lines and is outside the scope of this PR; suppression is now scoped to
those six modules + only the three relevant error codes (`assignment`,
`arg-type`, `return-value`) so the rest of cflibs no longer benefits from
silenced type errors.

Verification (Py3.12 venv with `pip install -e ".[dev,ci]"`, matches CI):

  ruff check cflibs/ tests/        →  All checks passed!
  black --check cflibs/            →  168 files would be left unchanged
  mypy cflibs/                     →  Success: no issues found in 168 source files
  pytest tests/ -m "not requires_db
    and not requires_bayesian
    and not requires_rust
    and not requires_jax
    and not slow"                  →  1330 passed, 2 skipped, 448 deselected

+++ cflibs/inversion/physics/self_absorption.py
@@ -1493,8 +1493,8 @@ class CurveOfGrowthAnalyzer:
         # Determine regime from slope
         regime = self._classify_regime(slope, r_squared)
 
-        # Estimate mean Doppler width
-        mean_wavelength = np.mean([ld.wavelength_nm for ld in lines_data])
+        # Estimate mean Doppler width. np.mean returns floating[Any]; coerce.
+        mean_wavelength = float(np.mean([ld.wavelength_nm for ld in lines_data]))
         doppler_width = self._compute_doppler_width(mean_wavelength)
 
         # Estimate damping parameter if not provided
